+++ src/task/JoinAndReceiveIndexTask.ts
@@ -21,7 +21,6 @@ import BaseTask from './BaseTask';
 export default class JoinAndReceiveIndexTask extends BaseTask {
   protected taskName = 'JoinAndReceiveIndexTask';
   private taskCanceler: TaskCanceler | null = null;
-  private maxVideos = 16;
 
   constructor(private context: AudioVideoControllerState) {
     super(context.logger);
@@ -99,7 +98,12 @@ export default class JoinAndReceiveIndexTask extends BaseTask {
       const interceptor = new IndexFrameInterceptor(this.context.signalingClient);
       this.context.signalingClient.registerObserver(interceptor);
       this.taskCanceler = interceptor;
-      this.context.signalingClient.join(new SignalingClientJoin(this.maxVideos, true));
+      this.context.signalingClient.join(
+        new SignalingClientJoin(
+          context.meetingSessionConfiguration.maxVideoStreamsSubscribeTo,
+          true
+        )
+      );
     });
     this.context.logger.info(`received first index ${JSON.stringify(indexFrame)}`);
     this.context.indexFrame = indexFrame;