+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update meeting readiness checker demo app to create meeting after the checker starts
 - Alter the versioning script to require less ritual
 - Correct TypeScript build to generate correct artifacts in `build/`
+- Make maxVideos in JoinFrame public configurable
 
 ### Removed